requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
django==3.1
pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
